switch to nasic_cstring.hpp instead of fwd header added failed expression to the system error system_error is caught and reported in cpp_main.ipp 64 bit compatibility issue in debug.ipp fixed missing headers in algorithm.hpp added
[SVN r41343]

+++ include/boost/test/impl/debug.ipp
@@ -315,7 +315,7 @@ prepare_window_title( dbg_startup_info const& dsi )
     else
         ++it;
 
-    ::snprintf( title_str, sizeof(title_str), "%*s %ld", dsi.binary_path.end()-it, it, dsi.pid );
+    ::snprintf( title_str, sizeof(title_str), "%*s %ld", (int)(dsi.binary_path.end()-it), it, dsi.pid );
 
     return title_str;
 }
